backend/backend-submit/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.10
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
environ==1.0
idna==3.6
kombu==5.3.5
pillow==10.2.0
prompt-toolkit==3.0.43
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
